Trash everything again
I'm working on an idea that makes everything look more OOP without actually changing how anything works. Instead of doing this:

    multiply (add 1 2) 3

You do this:

    1 :plus 2 :times 3

The idea is that it looks kind of object oriented. To wit:

    1.plus(2).times(3)

That's more or less what Ruby desugars into. (Of course for this particular example you'd just use operators.)